backend/requirements.txt
Django==5.0.3
psycopg[binary]==3.1.18
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pdfplumber==0.11.4
